requirements.txt
google-adk==0.1.0
litellm==1.66.1